Reorganize matching and more connectivity functions
